Fix typechecker bug
In the old version using Ty.t for types we represented an
unrefined base as a type refined with the trivial refinement
v[true]. With Ty_surface.t and Ty_template.t the refinement
is optional and so we instead set it to None. The helper
function unrefined_base wasn't up to date with this convention
though, so the type comparison in apply_types was failing.